requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LLM Client (Groq)
groq>=0.4.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data Processing (using versions with pre-built wheels)
pandas>=2.0.0
numpy>=1.24.0

# Vector Search
faiss-cpu==1.7.4

# ML/Embeddings (torch will auto-install compatible numpy)
sentence-transformers==2.3.1

# Validation & Config
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
jinja2==3.1.3
